visa_out_reports/outgoing/codes/requirements.txt
pandas
mysql-connector-python
watchdog
pdfplumber
openpyxl
flask
